Osp node reg update (#32)
* Update locators for new OSP node registion

QCI merged the pages for auto-detection and manual entry of nodes. The
id's for elements need to be updated accordingly

* Update locators for new OSP node registion

QCI merged the pages for auto-detection and manual entry of nodes. The
id's for elements need to be updated accordingly

* Update locators for new OSP node registion

QCI merged the pages for auto-detection and manual entry of nodes. The
id's for elements need to be updated accordingly

* Update locators for new OSP node registion

QCI merged the pages for auto-detection and manual entry of nodes. The
id's for elements need to be updated accordingly